+++ activerecord/test/cases/column_test.rb
@@ -118,6 +118,12 @@ module ActiveRecord
           end
         end
       end
+
+      def test_type_cast_two_digit_date
+        [Column.new("field", nil, "datetime"), Column.new("field", nil, "date")].each do |column|
+          assert_equal Date.parse("10/11/12"), column.type_cast("10/11/12")
+        end
+      end
     end
   end
 end