+++ activesupport/lib/active_support/core_ext/string/conversions.rb
@@ -5,7 +5,7 @@ class String
   # Converts a string to a Time value.
   # The +form+ can be either :utc or :local (default :local).
   #
-  # The time is parsed using Time.parse method.
+  # The time is parsed using Date.parse method.
   # If +form+ is :local, then the time is in the system timezone.
   # If the date part is missing then the current date is used and if
   # the time part is missing then it is assumed to be 00:00:00.
@@ -16,8 +16,9 @@ class String
   #   "2012-12-13T06:12".to_time         # => 2012-12-13 06:12:00 +0100
   #   "2012-12-13T06:12".to_time(:utc)   # => 2012-12-13 05:12:00 UTC
   #   "12/13/2012".to_time               # => ArgumentError: argument out of range
+  #   "13/12/4".to_time                  # => 2013-12-04 00:00:00 +0100
   def to_time(form = :local)
-    parts = Date._parse(self, false)
+    parts = Date._parse(self, true)
     return if parts.empty?
 
     now = Time.now
@@ -40,8 +41,9 @@ class String
   #   "01/01/2012".to_date # => Sun, 01 Jan 2012
   #   "2012-12-13".to_date # => Thu, 13 Dec 2012
   #   "12/13/2012".to_date # => ArgumentError: invalid date
+  #   "13/12/4".to_date    # => Wed, 04 Dec 2013
   def to_date
-    ::Date.parse(self, false) unless blank?
+    ::Date.parse(self, true) unless blank?
   end
 
   # Converts a string to a DateTime value.
@@ -50,7 +52,8 @@ class String
   #   "01/01/2012 23:59:59".to_datetime # => Sun, 01 Jan 2012 23:59:59 +0000
   #   "2012-12-13 12:50".to_datetime    # => Thu, 13 Dec 2012 12:50:00 +0000
   #   "12/13/2012".to_datetime          # => ArgumentError: invalid date
+  #   "13/12/4".to_datetime             # => Wed, 04 Dec 2013 00:00:00 +0000
   def to_datetime
-    ::DateTime.parse(self, false) unless blank?
+    ::DateTime.parse(self, true) unless blank?
   end
 end